% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{add_temporal}
\alias{add_temporal}
\title{Add temporal data to a nontemporal adjacency df}
\usage{
add_temporal(adj, t_vec = NULL)
}
\arguments{
\item{adj}{the nontemporal df to which we add a temporal column}

\item{t_vec}{a list of the timestamps to add.}
}
\description{
Add temporal data to a nontemporal adjacency df
}

