% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{adjacency_smoother}
\alias{adjacency_smoother}
\title{Construct an adjacency data frame for a grid_data data frame.}
\usage{
adjacency_smoother(grid_data, coords)
}
\arguments{
\item{grid_data}{a grid_data data frame}

\item{grid_data}{the grid_data data frame to smooth.}
}
\value{
a data frame giving the adjacencies between grid cells, suitable for
use as an edge-list to construct an igraph graph.
Smooth a grid_data data frame by constructing a new column of proportions
computed on the sum of a cell and each of its neighbors.

a new grid_data object with smoothed proportions
}
\description{
Construct an adjacency data frame for a grid_data data frame.
}

