% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractwise_clustering.R
\name{affinity_matrix}
\alias{affinity_matrix}
\title{Construct an affinity matrix for a graph g based on the dist edge attribute}
\usage{
affinity_matrix(g, sigma = 1)
}
\arguments{
\item{g}{an igraph::graph with a dist edge attribute/}

\item{sigma}{the multiplicative constant in the quadratic exponential
affinity kernel}
}
\value{
a matrix containing the affinities computed with the quadratic exponential
affinity kernel
}
\description{
Construct an affinity matrix for a graph g based on the dist edge attribute
}

