% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergences.R
\name{agg_loss}
\alias{agg_loss}
\title{Compute the agglomeration loss associated with combining two sets of counts
under a user-specified divergence}
\usage{
agg_loss(n, m, marginal = n + m, div = DKL)
}
\arguments{
\item{n}{the first set of counts to agglomerate}

\item{m}{the second set of counts to agglomerate}

\item{marginal}{the global count set}

\item{agg_div}{the divergence function under which to compute loss}

\item{...}{additional arguments passed to div}
}
\value{
a nonnegative scalar
}
\description{
Compute the agglomeration loss associated with combining two sets of counts
under a user-specified divergence
}

