% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{compute_derivatives}
\alias{compute_derivatives}
\title{Compute numerical derivatives from data using weighted linear regression
on each tract's 1-ego network.}
\usage{
compute_derivatives(adj, r_sigma = 1, s_sigma = 1, smooth = F)
}
\arguments{
\item{adj}{the adjacency matrix on which to compute. Must include centroids.}

\item{r_sigma}{geographic bandwidth of the regression weight decay}

\item{s_sigma}{if smooth = TRUE, the bandwidth of a geographic rbf smoother}

\item{smooth}{boolean should the derivatives be computed on a geographically smoothed demographic distribution?}
}
\value{
a tract-level tibble with a column D_alpha giving the derivatives of
demographic frequencies with respect to spatial coordinates.
}
\description{
Compute numerical derivatives from data using weighted linear regression
on each tract's 1-ego network.
}

