% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{compute_derivs}
\alias{compute_derivs}
\title{Compute numerical first derivatives on a grid_data data frame with a
user-specified difference method.}
\usage{
compute_derivs(grid_data, difference_methods = NULL)
}
\arguments{
\item{grid_data, }{a data frame giving counts on a grid.}

\item{difference_method, }{one of c('central', 'forward', 'backward')}
}
\value{
grid_data with new columns for derivatives, one for each grid
coordinate
}
\description{
Compute numerical first derivatives on a grid_data data frame with a
user-specified difference method.
}

