% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{compute_geometry}
\alias{compute_geometry}
\title{Compute local Riemannian metrics from tracts and a demographics data frame.
The user may specify the grid resolution, the type of divergence used, and
the numerical difference method employed}
\usage{
compute_geometry(tracts, data, resolution, divergence = "KL",
  difference_methods = NULL)
}
\arguments{
\item{tracts}{an SPDF}

\item{data}{a data frame consisting at least three columns:
- tract, the key relating the data frame to the tracts SPDF
- group, the group labels
- n a count for each group label in each tract
- (Optional) additional nonspatial coordinates, such as a time coordinate}

\item{resolution}{the spatial resolution of the grid on which to compute, in
km.}

\item{divergence}{one of c('KL', 'euc', 'cum_euc') specifying the divergence
used to compare distributions}

\item{difference_method}{one of c('central', 'forward', 'backward'),
specifying the computation of numerical first derivatives}
}
\value{
grid_data data frame with the the Riemannian metric tensor in local
coordinates for each grid cell.
}
\description{
Compute local Riemannian metrics from tracts and a demographics data frame.
The user may specify the grid resolution, the type of divergence used, and
the numerical difference method employed
}

