% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{compute_hessian}
\alias{compute_hessian}
\title{Compute the hessian (or Riemannian metric) at each tract.}
\usage{
compute_hessian(adj, hessian = DKL_)
}
\arguments{
\item{adj}{an adjacency tibble in which derivatives have already been computed}

\item{hessian}{one of `c(DKL_, euc_, cum_euc_)`}
}
\value{
a tibble with a column local_H containing the Hessian
(or Riemannian metric) in local coordinates.
}
\description{
Compute the hessian (or Riemannian metric) at each tract.
}

