% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R, R/tract_wise_geometry.R
\name{compute_metric}
\alias{compute_metric}
\title{Compute the Riemannian metric in local coordinates on a grid_data object}
\usage{
compute_metric(tracts, data, km = T, r_sigma = 100, s_sigma = 1,
  smooth = F, hessian = euc_)

compute_metric(tracts, data, km = T, r_sigma = 100, s_sigma = 1,
  smooth = F, hessian = euc_)
}
\arguments{
\item{grid_data}{a grid_data object}

\item{divergence}{the divergence function chosen to induce a metric in
information space. One of c('KL', 'euc', 'cum_euc').}

\item{coord_names}{the names of the coordinate columns in grid_data}
}
\value{
a new grid_data object with hessians (Riemannian metrics) computed
in local coordinates
}
\description{
Compute the Riemannian metric in local coordinates on a grid_data object

Compute the metric df
}

