% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbf_differentiation.R
\name{compute_metric_rbf}
\alias{compute_metric_rbf}
\title{Needs to handle time variables gracefully, and figure out problem
with DKL_}
\usage{
compute_metric_rbf(tracts, data, sigma = c(1, 1, 10), hessian = euc_)
}
\description{
Needs to handle time variables gracefully, and figure out problem
with DKL_
}

