% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_wise_geometry.R
\name{coords_df}
\alias{coords_df}
\title{Construct a tibble of centroid coordinates}
\usage{
coords_df(tracts, km = FALSE, ...)
}
\arguments{
\item{tracts}{the spdf whose coordinates we wish to compute.}

\item{...}{additional parameters passed to id_lookup}
}
\description{
Construct a tibble of centroid coordinates
}

