% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{create_test_data}
\alias{create_test_data}
\title{Generate some random data with correct structure for testing functions}
\usage{
create_test_data(n, d, K)
}
\arguments{
\item{the}{number of locations x}

\item{the}{spatial dimension}

\item{the}{number of characteristic distributions to use in modeling}
}
\value{
list a list of locations x, means Mu, covariance matrices Sigma, and representatives Q
}
\description{
Generate some random data with correct structure for testing functions
}

