% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detroit_race.R
\docType{data}
\name{detroit_race}
\alias{detroit_race}
\title{Tract-level racial demographic counts from 1970-2010 from the US Census.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15275 rows and 4 columns.}
\source{
\url{https://s4.ad.brown.edu/Projects/Diversity/Researcher/LTBDDload/Default.aspx}
}
\usage{
detroit_race
}
\description{
Tract-level racial demographic counts from 1970-2010 from the US Census.
}
\keyword{datasets}

