% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_measures.R
\name{DKL}
\alias{DKL}
\title{Find the Kullback-Leibler divergence of two empirical distributions.}
\usage{
DKL(p, q, check = FALSE)
}
\arguments{
\item{p}{the 'true' distribution}

\item{q}{the estimated distribution}

\item{check}{whether to check that p and q are valid probability distributions}
}
\value{
numeric the KL divergence between p and q
}
\description{
Find the Kullback-Leibler divergence of two empirical distributions.
}

