% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergences.R
\name{DKL_}
\alias{DKL_}
\title{Compute the Hessian of a Bregman divergence generator at a point}
\usage{
DKL_(base)
}
\arguments{
\item{base}{the base point at which to calculate the Hessian}
}
\description{
Compute the Hessian of a Bregman divergence generator at a point
}

