% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/hierarchical_clustering.R
\name{empty_hclust}
\alias{empty_hclust}
\title{Initialize an empty hclust object to be populated by the loop in info_clust}
\usage{
empty_hclust(n)

empty_hclust(n)
}
\arguments{
\item{n}{the number of leaves}

\item{n}{the number of objects to be clustered}
}
\value{
a a list with entries appropriate for conversion to an hclust object.
}
\description{
Initialize an empty hclust object to be populated by the loop in info_clust

Construct an empty object of class 'hclust'
}

