% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{empty_merge_lookup}
\alias{empty_merge_lookup}
\title{Create an empty merge lookup for use populating a$merge}
\usage{
empty_merge_lookup(n)
}
\arguments{
\item{n}{the number of leaves}
}
\value{
merge_lookup a list
}
\description{
Create an empty merge lookup for use populating a$merge
}

