% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{estimate}
\alias{estimate}
\title{Find the spatially-structured estimate at a location x}
\usage{
estimate(x, Q, Mu, Sigma, C = 1)
}
\arguments{
\item{x}{vector the location}

\item{Q}{matrix the matrix of representative distributions}

\item{Mu}{list a list of means}

\item{Sigma}{list a list of covariance matrices}

\item{C}{vector the vector of scale coefficients}
}
\value{
vector the estimates at x
}
\description{
Find the spatially-structured estimate at a location x
}

