% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractwise_clustering.R
\name{generalized_laplacian_matrix}
\alias{generalized_laplacian_matrix}
\title{Construct a generalized laplacian matrix}
\usage{
generalized_laplacian_matrix(input, ...)
}
\arguments{
\item{input}{either an affinity matrix or an igraph::graph. If
class(input) == 'igraph', then an affinity matrix is computed and used.
Otherwise, it is assumed that the input is an affinity matrix.}

\item{...}{additional arguments passed to affinity_matrix if
class(input) == 'igraph'.}
}
\value{
a generalized laplacian matrix suitable for spectral analysis.
}
\description{
Construct a generalized laplacian matrix
}
\details{
This function involves a matrix inversion and is therefore fairly
expensive for large graphs.
}

