% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_census_data}
\alias{get_census_data}
\title{Make a set of joined data based on specified census information at the block level.}
\usage{
get_census_data(state, table_nums = c("B15003", "B03002", "C24010"),
  lookup_table = NULL, end_year = 2014, span = 5)
}
\arguments{
\item{state}{the state for which to obtain data.}

\item{end_year}{the final year of the data set to use}

\item{span}{the number of years the data should extend back}

\item{counties}{a vector of ints reflecting counties to use.}

\item{table_name}{character name of the census table t use}
}
\value{
merged_df a data frame including both geographical and data fields.
}
\description{
Make a set of joined data based on specified census information at the block level.
}

