% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_centroids}
\alias{get_centroids}
\title{Get the centroids for a bunch of spatial data.}
\usage{
get_centroids(data)
}
\arguments{
\item{data}{the SpatialPolygon data for which to compute centroids}
}
\value{
a matrix of centroid coordinates
}
\description{
Get the centroids for a bunch of spatial data.
}

