% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table}
\alias{get_table}
\title{Get matching df for a bunch of spatial data}
\usage{
get_table(data)
}
\arguments{
\item{data}{the SpatialPolygon data for which to compute centroids}
}
\value{
df the complete set of data corresponding to the SpatialPolygons, keyed by ID to match with them.
}
\description{
Get matching df for a bunch of spatial data
}

