% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{grad_f}
\alias{grad_f}
\title{Find the gradient of the part of the Kullback-Leibler divergence that depends on f}
\usage{
grad_f(p, q)
}
\arguments{
\item{p}{the 'true' distribution}

\item{q}{the estimated distribution}
}
\value{
p dot log(q)
}
\description{
Find the gradient of the part of the Kullback-Leibler divergence that depends on f
}

