% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_aggregation.R
\name{grid_aggregate}
\alias{grid_aggregate}
\title{Construct grid_data from tracts and data, given a fixed grid resolution. This
is a wrapper for make_grid, make_grid_tract, make_coordinates, and
make_grid_data.}
\usage{
grid_aggregate(tracts, resolution, data)
}
\arguments{
\item{tracts}{an SPDF}

\item{resolution}{the resolution of the grid at which to aggregate, in km}

\item{data}{a data frame consisting at least three columns:
- tract, the key relating the data frame to the tracts SPDF
- group, the group labels
- n a count for each group label in each tract
- (Optional) additional nonspatial coordinates, such as a time coordinate}
}
\value{
a list containing grid_tract and grid_Data.
}
\description{
Construct grid_data from tracts and data, given a fixed grid resolution. This
is a wrapper for make_grid, make_grid_tract, make_coordinates, and
make_grid_data.
}

