% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_measures.R
\name{H}
\alias{H}
\title{Find the entropy of a distribution}
\usage{
H(p)
}
\arguments{
\item{p}{a normalized vector of nonnegative probabilities.}
}
\description{
Find the entropy of a distribution
}

