% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_measures.R
\name{H_B}
\alias{H_B}
\title{Compute the binary entropy function for a fixed proportion.}
\usage{
H_B(p)
}
\arguments{
\item{p}{the proportion}
}
\description{
Compute the binary entropy function for a fixed proportion.
}

