% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{h_clust}
\alias{h_clust}
\title{Conduct hierarchical clustering of either an sf geography with demographic data or an information graph}
\usage{
h_clust(input, data)
}
\arguments{
\item{input}{either an sf data.frame of polygons or an information graph as constructed by construct_information_graph}

\item{data}{if input is an sf data.frame, the demographic data for each geographic unit}
}
\value{
a, an hclust object
}
\description{
Conduct hierarchical clustering of either an sf geography with demographic data or an information graph
}

