% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{id_lookup}
\alias{id_lookup}
\title{Construct a lookup data frame from an SPDF object}
\usage{
id_lookup(tracts, key_col = "GEOID")
}
\arguments{
\item{tracts}{the object whose keys to retrieve}

\item{key_col}{the column of tracts@data containing the identifier.}
}
\value{
df a two-column tibble in which the first column is the id and
the second is the lookup identifier.
}
\description{
Construct a lookup data frame from an SPDF object
}

