% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{info_clust}
\alias{info_clust}
\title{Compute constrained, greedy information-maximizing agglomerative clustering on a contingency table.}
\usage{
info_clust(df, constraint)
}
\arguments{
\item{df}{a data frame in which each row is a separate observation and each column a category.}

\item{constraint}{a matrix of constraints. Let constraint be a matrix of TRUE to do unconstrained clustering.}
}
\value{
a an object of class hclust
}
\description{
Compute constrained, greedy information-maximizing agglomerative clustering on a contingency table.
}

