% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{info_loss}
\alias{info_loss}
\title{Compute the information loss associated with merging two rows of a data frame.}
\usage{
info_loss(df, i, j)
}
\arguments{
\item{df}{a data frame in which each column represents a category and each row the observed counts in that category.}

\item{i}{the first row to merge}

\item{i}{the second row to merge}
}
\description{
Compute the information loss associated with merging two rows of a data frame.
}

