% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{information_distances}
\alias{information_distances}
\title{Early stage: pass hessians as names, pass DJS as a function.}
\usage{
information_distances(adj, divergence, ...)
}
\description{
Early stage: pass hessians as names, pass DJS as a function.
}

