% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_aggregation.R
\name{make_coordinates}
\alias{make_coordinates}
\title{Construct a data frame of coordinates for a grid SPDF}
\usage{
make_coordinates(grid, dims, resolution, nonspatial_coords = NULL)
}
\arguments{
\item{grid}{a square grid SpatialPolygonsDataFrame}

\item{dims}{the dimensions of grid}

\item{resolution}{the spatial resolution of the grid, i.e. the units
corresponding to a single side of a cell.}

\item{nonspatial_coords}{a list of nonspatial coordinates. The standard
use case is the addition of time variables.}
}
\value{
a data frame consisting of a rectangular coordinates for a grid
SPDF.
}
\description{
Construct a data frame of coordinates for a grid SPDF
}

