% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R, R/grid_aggregation.R
\name{make_grid}
\alias{make_grid}
\title{Construct a spatial grid over a SpatialPolygonsDataFrame}
\usage{
make_grid(tracts, resolution)

make_grid(tracts, resolution)
}
\arguments{
\item{tracts}{the spdf}

\item{resolution}{the radius of the grid to produce}

\item{tracts}{a SpatialPolygonsDataFrame}

\item{resolution}{the length in km of the grid cells}
}
\value{
a list containing the grid as a SpatialPolygonsDataFrame, the
resolution, and the grid dimensions.
}
\description{
Construct a spatial grid over a SpatialPolygonsDataFrame

Construct a square grid over input tracts with specified
spatial resolution.
}

