% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_aggregation.R
\name{make_grid_data}
\alias{make_grid_data}
\title{Construct a data frame consisting of data originally keyed to a tracts
 object, re-keyed to the grid. The grid and tracts must be related by
 grid_tract.}
\usage{
make_grid_data(grid_tract, data, coords)
}
\arguments{
\item{grid_tract}{a data frame, produced by `make_grid_tract()`, relating
a grid SPDF to a tracts SPDF}

\item{data}{the data frame keyed to the tract object.}

\item{coords}{the coordinates of the grid, as produced by
`make_coordinates()`}
}
\value{
a data frame giving counts from data, re-keyed to grid, as well as
spatial (rectangular) coordinates of the grid variables.
}
\description{
Construct a data frame consisting of data originally keyed to a tracts
 object, re-keyed to the grid. The grid and tracts must be related by
 grid_tract.
}

