% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_aggregation.R
\name{make_grid_tract}
\alias{make_grid_tract}
\title{Construct a data frame that relates the grid cells to the original tracts.
There is one row for each instance of a tract overlapping a grid cell.
This operation can be quite computationally expensive.}
\usage{
make_grid_tract(grid, tracts)
}
\arguments{
\item{grid}{a SpatialPolygonsDataFrame}

\item{tracts}{a SpatialPolygonsDataFrame}
}
\value{
a data frame relating grid to tracts.
}
\description{
Construct a data frame that relates the grid cells to the original tracts.
There is one row for each instance of a tract overlapping a grid cell.
This operation can be quite computationally expensive.
}

