% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{make_merge_list}
\alias{make_merge_list}
\title{Compute the merge list for a single pass of agglomerative hierarchical clustering.
The result is a df in which each row corresponds to a pair of nodes
and a pair is present if each node is the nearest neighbor
of the other. Also includes stage information (labels)
and the loss associated with agglomerating the two nodes.}
\usage{
make_merge_list(h, divergence, current_stage = 0)
}
\arguments{
\item{h}{the graph to cluster. Must have "stage" information for the nodes}

\item{divergence}{the divergence function used to compute distances between nodes}

\item{current_stage}{the number of agglomerations performed so far}
}
\value{
a data frame giving the merges and loss associated with each one
}
\description{
Compute the merge list for a single pass of agglomerative hierarchical clustering.
The result is a df in which each row corresponds to a pair of nodes
and a pair is present if each node is the nearest neighbor
of the other. Also includes stage information (labels)
and the loss associated with agglomerating the two nodes.
}

