% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{merge_graph}
\alias{merge_graph}
\title{Merge nodes in a graph based on an input merge list}
\usage{
merge_graph(h, merge_list)
}
\arguments{
\item{h}{the graph to merge}

\item{merge_list}{a data frame giving the merges and
losses associated with each merge}
}
\value{
a graph h with specified nodes merged
}
\description{
Merge nodes in a graph based on an input merge list
}

