% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_measures.R
\name{mutual_info}
\alias{mutual_info}
\title{Compute the mutual information of a data frame or matrix of crosstabs}
\usage{
mutual_info(input)
}
\arguments{
\item{input}{a data frame or matrix representing a crosstab}

\item{drop_threshold}{numerical keyword to KL divergence avoiding 0*0 multiplication. Should not be modified.}
}
\description{
Compute the mutual information of a data frame or matrix of crosstabs
}

