% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergences.R
\name{NA_multiply}
\alias{NA_multiply}
\title{Quadratic form multiplication with NA and zero logic. The modified logic is:
0 * Inf = 0
0 * NA = 0}
\usage{
NA_multiply(delta, H)
}
\arguments{
\item{delta}{a vector or matrix}

\item{H}{a square matrix}
}
\value{
a matrix, the product t(delta) %*% .H %*% delta
}
\description{
Quadratic form multiplication with NA and zero logic. The modified logic is:
0 * Inf = 0
0 * NA = 0
}

