% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{normal_vec}
\alias{normal_vec}
\title{Compute a vector of values of different normal densities at a point x}
\usage{
normal_vec(x, Mu, Sigma)
}
\arguments{
\item{x}{vector the location}

\item{Mu}{a list of means}

\item{Sigma}{a list of covariance matrices}
}
\value{
a vector of densities at x corresponding to the different densities
}
\description{
Compute a vector of values of different normal densities at a point x
}

