% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_clustering.R
\name{normalized_laplacian}
\alias{normalized_laplacian}
\title{Construct a generalized laplacian matrix}
\usage{
normalized_laplacian(A)
}
\arguments{
\item{input}{either an affinity matrix or an igraph::graph. If
class(input) == 'igraph', then an affinity matrix is computed and used.
Otherwise, it is assumed that the input is an affinity matrix.}

\item{...}{additional arguments passed to affinity_matrix if
class(input) == 'igraph'.}
}
\value{
a generalized laplacian matrix suitable for spectral analysis.
}
\description{
Construct a generalized laplacian matrix
}
\details{
This function involves a matrix inversion and is therefore fairly
expensive for large graphs.
}

