% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergences.R
\name{quad}
\alias{quad}
\title{Quadratic approximation to a Bregman divergence}
\usage{
quad(n, m, base = m, h = euc_)
}
\arguments{
\item{n}{the first point}

\item{m}{the second point}

\item{base}{the point at which to compute the Hessian}

\item{h}{the hessian function. Will be passed base as argument}
}
\value{
the quadratic approximation of the Bregman divergence of which h is
the hessian.
}
\description{
Quadratic approximation to a Bregman divergence
}

