% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_measures.R
\name{simplex_check}
\alias{simplex_check}
\title{Check whether a vector is an element of the probability simplex}
\usage{
simplex_check(p, allow_zero = TRUE)
}
\arguments{
\item{p}{vector the vector to check}

\item{allow_zero}{boolean whether to allow elements with zero entries (boundary of simplex)}
}
\value{
boolean whether p is a valid probability distribution
}
\description{
Check whether a vector is an element of the probability simplex
}

