% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_measures.R
\name{simplex_normalize}
\alias{simplex_normalize}
\title{Normalize a nonnegative vector so that it lies in the probability simplex}
\usage{
simplex_normalize(p)
}
\arguments{
\item{p}{vector the vector}
}
\description{
Normalize a nonnegative vector so that it lies in the probability simplex
}

