% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{update_constraint}
\alias{update_constraint}
\title{Update the constraint matrix using the current merge list. Note that, if no merges are possible, then the constraint matrix inserts a TRUE at positions [1,2] and [2,1]. In the geographical context, this scenario corresponds to disconnected components on the map, and the constraint matrix declares two of them to be connected arbitrarily.}
\usage{
update_constraint(constraint, merge_list)
}
\arguments{
\item{constraint}{the current state of the constraint matrix.}

\item{merge_list}{the list of observations to be merged.}
}
\value{
matrix the updated constraint matrix
}
\description{
Update the constraint matrix using the current merge list. Note that, if no merges are possible, then the constraint matrix inserts a TRUE at positions [1,2] and [2,1]. In the geographical context, this scenario corresponds to disconnected components on the map, and the constraint matrix declares two of them to be connected arbitrarily.
}

