% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{update_df}
\alias{update_df}
\title{Update the data frame of observations using the current merge list.}
\usage{
update_df(df, merge_list)
}
\arguments{
\item{df}{the current state of the df}

\item{merge_list}{the list of observations to be merged.}
}
\value{
df the updated data frame
}
\description{
Update the data frame of observations using the current merge list.
}

