% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{adj_with_coords}
\alias{adj_with_coords}
\title{Construct a (nested) tibble with tract-level demographics and coordinates}
\usage{
adj_with_coords(adj, tracts, km = FALSE)
}
\arguments{
\item{tracts}{the sf data.frame supplying geographic information}

\item{km}{boolean should the centroids be converted so that the euclidean distances between them are approximately measurable in km?}

\item{data}{the tibble containing the demographics for the tracts.}
}
\value{
a nested tibble containing tract-level demographics and coordinates
as vectors in list-columns.
}
\description{
Construct a (nested) tibble with tract-level demographics and coordinates
}

