% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_wise_geometry.R
\name{compute_derivatives}
\alias{compute_derivatives}
\title{Compute numerical derivatives from data using weighted linear regression
on each tract's 1-ego network.}
\usage{
compute_derivatives(adj, r_sigma = 1, s_sigma = 1, smooth = F, ...)
}
\arguments{
\item{tracts, }{the spdf}

\item{data, }{a tibble containing the demographic data}

\item{sigma}{the multiplicative constant in the square exponential kernel.}
}
\value{
a tract-level tibble with a column D_alpha giving the derivatives of
demographic frequencies with respect to spatial coordinates.
}
\description{
Compute numerical derivatives from data using weighted linear regression
on each tract's 1-ego network.
}

