% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tract_wise_geometry.R
\name{compute_hessian}
\alias{compute_hessian}
\title{Compute the hessian (or Riemannian metric) at each tract.}
\usage{
compute_hessian(adj, divergence = "KL")
}
\arguments{
\item{divergence}{one of c('KL', 'euc', 'cum_euc')}

\item{tracts}{an spdf}

\item{data}{a tibble containing demographic data}

\item{sigma}{multiplicative constant on the square exponential kernel weights}
}
\value{
a tibble with a column local_H containing the Hessian
(or Riemannian metric) in local coordinates.
}
\description{
Compute the hessian (or Riemannian metric) at each tract.
}

