% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{compute_metric}
\alias{compute_metric}
\title{Compute the pullback metric tensor from an sf data frame and an accompanying set of
demographic data.}
\usage{
compute_metric(tracts, data, km = T, r_sigma = 100, s_sigma = 1,
  smooth = F, hessian = euc_)
}
\arguments{
\item{tracts}{the sf data.frame supplying geographic information}

\item{data}{the tibble supplying demographic data}

\item{km}{whether to compute rectangular distances in km}

\item{r_sigma}{regression smoother for derivative computation}

\item{s_sigma}{geographic smoother for derivative computation}

\item{smooth}{boolean should a geographic smoother be used for derivatives?}

\item{hessian}{the hessian function to be used, one of c(DKL_, euc_, cum_euc_)}
}
\value{
a tibble in which each row has a list-column containing the pullback metric tensor stored as a matrix.
}
\description{
Compute the pullback metric tensor from an sf data frame and an accompanying set of
demographic data.
}

