% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{coords_df}
\alias{coords_df}
\title{Construct a tibble of centroid coordinates from an sf data.frame containing polygons in its geometry column.}
\usage{
coords_df(tracts, km = FALSE, ...)
}
\arguments{
\item{tracts}{the sf data.frame whose coordinates we wish to compute.}

\item{km}{boolean should the centroids be converted so that the euclidean distances between them are approximately measurable in km?}
}
\description{
Construct a tibble of centroid coordinates from an sf data.frame containing polygons in its geometry column.
}

