% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergences.R
\name{DKL}
\alias{DKL}
\title{Find the Kullback-Leibler divergence of two empirical distributions.}
\usage{
DKL(n, m)
}
\arguments{
\item{p}{the 'true' distribution}

\item{q}{the estimated distribution}
}
\value{
numeric the divergence between p and q
}
\description{
Find the Kullback-Leibler divergence of two empirical distributions.
}

