% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{empty_hclust}
\alias{empty_hclust}
\title{Construct an empty object of class 'hclust'}
\usage{
empty_hclust(n)
}
\arguments{
\item{n}{the number of objects to be clustered}
}
\description{
Construct an empty object of class 'hclust'
}

