% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{information_distances}
\alias{information_distances}
\title{We are assuming a divergence function, defined by the user,
that takes in n_1 and n_2 and returns a number.}
\usage{
information_distances(adj, divergence)
}
\description{
We are assuming a divergence function, defined by the user,
that takes in n_1 and n_2 and returns a number.
}

