% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{make_grid}
\alias{make_grid}
\title{Construct a spatial grid over a SpatialPolygonsDataFrame}
\usage{
make_grid(tracts, resolution)
}
\arguments{
\item{tracts}{the spdf}

\item{resolution}{the radius of the grid to produce}
}
\description{
Construct a spatial grid over a SpatialPolygonsDataFrame
}

