% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{make_merge_list}
\alias{make_merge_list}
\title{Make the list of merges for each outer iteration}
\usage{
make_merge_list(df, constraint)
}
\arguments{
\item{df}{the dataframe at this stage of the algorithm.}

\item{constraint}{the constraint at this stage of the algorithm}
}
\value{
merge_list a data frame with columns i and j (to be merged) and loss, the information loss associated with each.
}
\description{
Make the list of merges for each outer iteration
}

