% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{sort_merges}
\alias{sort_merges}
\title{Sort a merge list so that the corresponding hierarchical clustering is indeed greedy.}
\usage{
sort_merges(merges)
}
\arguments{
\item{merges}{the merge list to sort}
}
\value{
the merge list sorted, with appropriately relabeled indices.
}
\description{
Sort a merge list so that the corresponding hierarchical clustering is indeed greedy.
}

