% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/hierarchical_clustering.R
\name{sort_merges}
\alias{sort_merges}
\title{Sort the full list of merges produced in the main loop to ensure the greediness of the algorithm. While this version of agglomerative clustering is fast, the merge list generated is not on its own greedy. We can greedify the algorithm using the sorting procedure below, ensuring that each entry of a$height is minimal given the previous merges. *Sorting the list does not change the cluster topology*, just the order of the merges in the tree.}
\usage{
sort_merges(merges)

sort_merges(merges)
}
\arguments{
\item{merges}{the data frame of merges}

\item{merges}{the merge list to sort}
}
\value{
sorted the sorted data frame of merges

the merge list sorted, with appropriately relabeled indices.
}
\description{
Sort the full list of merges produced in the main loop to ensure the greediness of the algorithm. While this version of agglomerative clustering is fast, the merge list generated is not on its own greedy. We can greedify the algorithm using the sorting procedure below, ensuring that each entry of a$height is minimal given the previous merges. *Sorting the list does not change the cluster topology*, just the order of the merges in the tree.

Sort a merge list so that the corresponding hierarchical clustering is indeed greedy.
}

