% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tractwise_clustering.R
\name{spectral_cluster}
\alias{spectral_cluster}
\title{Cluster a graph or affinity matrix.
Output is a kmeans object giving the clusters.}
\usage{
spectral_cluster(g, sigma = 1, k = 2, nreps = 100)
}
\description{
Cluster a graph or affinity matrix.
Output is a kmeans object giving the clusters.
}

